pyproject.toml
[project]
name = "dark-cv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "opencv-python-headless<5",
    "pillow<11",
    "retinex>=0.0.1",
    "typer>=0.16.0",
]

[tool.uv]
package = true

[project.scripts]
dark-cv = "dark_cv.main:app"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.1",
    "ty>=0.0.1a12",
]
